pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Augusto Balsanelli", email = "augustebalsanel@proton.me"},
]
dependencies = [
    "django>=4.2.3",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "setuptools>=68.0.0",
    "djangorestframework-simplejwt>=5.2.2",
    "requests>=2.31.0",
    "drf-spectacular>=0.26.4",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "drf-writable-nested>=0.7.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]
